bugfix
see the comments.

Also, I was planning on disabling the run-every-frame of $00cde8, because setting $71 to $0B causes the controls to be cleared every frame, so carrying sprites can cause the player to drop the sprite during a screen transition (mainly that the player will drop it and pick it back up).