failsafe added
This failsafe prevents the screen from escaping the level borders when the custom borders are enabled and the borders extends outside the level.

If you are wondering why are the branch and code swapped places for the patch Initally, I thought it was a good idea I could just run "vanilla" after the custom limit code as to act as a failsafe should the custom border goes outside the level. Too bad, this causes the screen to go past the custom limit (and vibrate) due to the fact that "vanilla" also contains code that the screen is set to follow the player. So running both custom and then vanilla is a bad idea.